+++ rootbox/cli/cmd_exec.py
@@ -1,15 +1,20 @@
 """  Create a new environment  """
 
-from typing import Optional
+from pathlib import Path
 
 import typer
+from typing_extensions import Annotated
 
 from ..enter import enter_process
 
 
 def exec(
-    pid: int,
-    no_shell: bool = typer.Option(False, "--no-sh"),
-    command: Optional[str] = typer.Argument("/bin/sh", help="Command to be run"),
+    command: Annotated[str, typer.Argument(help="Command to be run")] = None,
+    no_shell: bool = typer.Option(
+        False,
+        "--no-sh",
+    ),
 ):
+    config_dir = Path.home().joinpath(".rootbox")
+    pid = int(Path(config_dir, ".lastpid").read_text())
     enter_process(pid, no_shell, command)